# Refactor DocumentPaneProvider