--- a/packages/@sanity/desk-tool/src/panes/document/documentPanel/documentViews/FormView.tsx
+++ b/packages/@sanity/desk-tool/src/panes/document/documentPanel/documentViews/FormView.tsx
@@ -20,6 +20,7 @@ import {ObjectField, ObjectSchemaTypeWithOptions} from '@sanity/types'
 import {useDocumentPane} from '../../useDocumentPane'
 import {Delay} from '../../../../components/Delay'
 import {useConditionalToast} from './useConditionalToast'
+import {useEditState} from '@sanity/react-hooks'
 
 interface FormViewProps {
   granted: boolean
@@ -46,8 +47,6 @@ export function FormView(props: FormViewProps) {
   const {hidden, margins, granted} = props
   const {
     compareValue,
-    displayed: value,
-    editState,
     documentId,
     documentSchema,
     documentType,
@@ -55,10 +54,13 @@ export function FormView(props: FormViewProps) {
     handleChange,
     handleFocus,
     historyController,
+    initialValue,
     markers,
     ready,
     changesOpen,
   } = useDocumentPane()
+  const editState = useEditState(documentId, documentType)
+  const value = editState.draft || editState.published || initialValue
   const presence = useDocumentPresence(documentId, {ignoreLastActiveUpdates: true})
   const {revTime: rev} = historyController
   const [{filterField}, setFilterField] = useState<FormViewState>(INITIAL_STATE)