--- a/packages/@sanity/react-hooks/src/useEditState.ts
+++ b/packages/@sanity/react-hooks/src/useEditState.ts
@@ -4,10 +4,25 @@
 import type {EditStateFor} from '@sanity/base/_internal'
 import documentStore from 'part:@sanity/base/datastore/document'
 import {useMemoObservable} from 'react-rx'
+import {merge, timer} from 'rxjs'
+import {debounce, share, skip, take} from 'rxjs/operators'
 
-export function useEditState(publishedDocId: string, docTypeName: string): EditStateFor {
-  return useMemoObservable(() => documentStore.pair.editState(publishedDocId, docTypeName), [
-    publishedDocId,
-    docTypeName,
-  ]) as EditStateFor
+export function useEditState(
+  publishedDocId: string,
+  docTypeName: string,
+  priority: 'default' | 'low' = 'default'
+): EditStateFor {
+  return useMemoObservable(() => {
+    const base = documentStore.pair.editState(publishedDocId, docTypeName).pipe(share())
+    if (priority === 'low') {
+      return merge(
+        base.pipe(take(1)),
+        base.pipe(
+          skip(1),
+          debounce(() => timer(1000))
+        )
+      )
+    }
+    return documentStore.pair.editState(publishedDocId, docTypeName)
+  }, [publishedDocId, docTypeName, priority]) as EditStateFor
 }